pyproject.toml
[project]
name = "daily-brief-intel-bi"
version = "0.0.1"
description = "Local-first BI tool PR0 foundation"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.9.2",
    "fastapi>=0.110.0",
    "filelock>=3.12.0",
    "jinja2>=3.1.2",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=7.4.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
